+++ CHANGES.md
@@ -18,7 +18,12 @@ Search:
     Others:
 
 Interoperability
+
 -   Fixed /dashboard and /dataset?q=keyword redirects for migrating easily from CKAN sites
+
+Others:
+
+-   Changed terminology for data rating to `Linked Data Rating`
 
 ## 0.0.51